Move search code into own package
