+++ src/us/kbase/workspace/database/Workspace.java
@@ -44,9 +44,10 @@ import us.kbase.typedobj.idref.IdReferenceHandlerSetFactory.IdReferenceHandlerFa
 import us.kbase.typedobj.idref.IdReferenceType;
 import us.kbase.typedobj.idref.RemappedId;
 import us.kbase.workspace.database.ResourceUsageConfigurationBuilder.ResourceUsageConfiguration;
-import us.kbase.workspace.database.SearchReferenceDAG.ReferenceDAGTopologyProvider;
-import us.kbase.workspace.database.SearchReferenceDAG.ReferenceProviderException;
-import us.kbase.workspace.database.SearchReferenceDAG.ReferenceSearchMaximumSizeExceededException;
+import us.kbase.workspace.database.refsearch.SearchReferenceDAG;
+import us.kbase.workspace.database.refsearch.SearchReferenceDAG.ReferenceDAGTopologyProvider;
+import us.kbase.workspace.database.refsearch.SearchReferenceDAG.ReferenceProviderException;
+import us.kbase.workspace.database.refsearch.SearchReferenceDAG.ReferenceSearchMaximumSizeExceededException;
 import us.kbase.workspace.database.exceptions.CorruptWorkspaceDBException;
 import us.kbase.workspace.database.exceptions.InaccessibleObjectException;
 import us.kbase.workspace.database.exceptions.NoSuchObjectException;
